Move Prefix to a separate file/module
