+++ mountpoint-s3/tests/common/mod.rs
@@ -2,7 +2,8 @@ use aws_sdk_s3::types::ByteStream;
 use aws_sdk_s3::Region;
 use fuser::{FileAttr, FileType};
 use futures::executor::ThreadPool;
-use mountpoint_s3::fs::{DirectoryReplier, Prefix, ReadReplier};
+use mountpoint_s3::fs::{DirectoryReplier, ReadReplier};
+use mountpoint_s3::prefix::Prefix;
 use mountpoint_s3::{S3Filesystem, S3FilesystemConfig};
 use mountpoint_s3_client::mock_client::{MockClient, MockClientConfig};
 use mountpoint_s3_crt::common::rust_log_adapter::RustLogAdapter;
