+++ mountpoint-s3/tests/fuse_tests/mod.rs
@@ -6,8 +6,8 @@ mod readdir_test;
 mod write_test;
 
 use fuser::{BackgroundSession, MountOption, Session};
-use mountpoint_s3::fs::Prefix;
 use mountpoint_s3::fuse::S3FuseFilesystem;
+use mountpoint_s3::prefix::Prefix;
 use mountpoint_s3::S3FilesystemConfig;
 use tempfile::TempDir;
